pyproject.toml
[tool.poetry]
name = "customer-sentiment-hub"
version = "0.1.0"
description = "A modular system for analyzing customer sentiment in debt resolution reviews"
authors = ["Nedal Altiti <nedal.a@usclarity.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "customer_sentiment_hub", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.4.0"
langchain = "^0.1.0"  # Use a newer major version
langchain-google-vertexai = "^0.1.0"
google-cloud-aiplatform = "^1.36.0"
typer = "^0.9.0"
rich = "^13.6.0"
python-dotenv = "^1.0.0"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"
pylint = "^3.0.2"
pre-commit = "^3.4.0"

[tool.poetry.scripts]
sentiment-hub = "customer_sentiment_hub.cli.app:app"
sentiment-api = "customer_sentiment_hub.api.app:start"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
